* feat(exp randomization): shuffle experiments then sort by repetition
Good experiment design mandates that the result of experiments hold even if factors that we do not look at change. This ensures that whatever result we are getting is really just the result of the variables we look it.

One way to bring a bit of randomness into the exomat is by actively shuffling the execution of the runs of an experiment. This way it is ensured that the first experiment running on a "cold" machine is not always the same one.

The shuffling can arbitrarily mix the execution of different envs ( env2 -> env1 -> env0 -> env1 ...) but will keep the order of repetitions (So that for any single env, execution of repetition 4 came after repetition 3)
 
This fixes #7
---------

Co-authored-by: teto519f <tessa.todorowski@tu-dresden.de>